Make add return a clone of the center of mass
JavaScript passes by references. Because of this, the center of mass body
in the tree is the same as the body that was inserted into the tree. Since
the function was mutating the center of mass instead of cloning it, the original
body was changed also. This causes the mass of the body to infinitely get larger
with each iteration, and completely breaks the physics of the simulation. By
returning a clone, this problem is fixed.